fix(KEN-865): the record suite runs under any uid [no-changelog]
Two threads on #1872, both defects this branch introduced.

Three passages still explained the old cause. A replacement made with
privilege used to write its record into the privileged account's data
directory; it now writes no record at all. The mechanism each passage is
there for survives the change, so each is reworded rather than cut:
`refresh_the_replaced_bytes` still exists because a privileged run leaves
the record naming bytes that are gone.

The second thread is the suite itself. `an_unprivileged_run_still_records`
asserted `!acting_as_root()`, which fails by construction under uid 0, and
a root dev container is a normal way to run this repo. The larger half was
quieter: every case that reached a record writer through a public entry
point became a no-op under root. Eighteen failed and six more passed while
asserting against a record the guard had refused to write.

Cases that only need a record on disk now write it through the `_as` seam
with the identity spelled out, so what they assert no longer depends on
who started the runner. `record_installed` gains the same seam its two
neighbours already had.

The rest cannot be isolated. `crates/cli/tests/command_record.rs` drives
the built binary on purpose, and that binary reads its own uid exactly as
it does in the field; the `update` cases and one lookup case drive
production writes with no seam in front of them. Those say why they are
skipping and stop, rather than passing for a write that never happened.
The uid is read from the syscall and not from `acting_as_root`, so a build
whose guard answered a constant cannot talk them out of running: with the
guard pinned to `true`, fifteen of them fail rather than skip.

The wiring proof stays and now covers both arms.
`a_public_write_follows_this_process_uid` calls the public entry point and
holds its answer against `geteuid`, read at the test site rather than
through the function whose wiring is in question. A build answering `true`
always fails it on any unprivileged runner, and one answering `false`
always fails it under root. Neither uid is skipped.

`command_update/tests.rs` sat at exactly the 800-line cap, so it could not
take the lines this needed. It splits at the seam the record suite next
door already names: the lookup against a described machine, which needs no
file on disk to decide anything, moves to `command_update/tests/described.rs`
and leaves the cases that read real files behind. No case is added or
dropped by the move, and none of the moved ones is touched.

Counts on this worktree: kendex-core and kendex-cli 2136 pass, 0 fail;
kendex-app lib 74 pass, 0 fail. Under `unshare -Ur`, 2118 pass and 18
fail, all eighteen `chmod 0000` cases that root defeats because root
bypasses file permissions. None of the eighteen touches the record.
