+++ crates/app/Cargo.toml
@@ -43,6 +43,12 @@ tauri-plugin-updater = "2"
 png = "0.18"
 tempfile.workspace = true
 
+[target.'cfg(unix)'.dev-dependencies]
+# The effective uid, read straight from the syscall so a suite can say why
+# it can observe no record on a runner that is already root — a read the
+# guard under test cannot answer for. Already in the workspace under core.
+rustix.workspace = true
+
 [build-dependencies]
 tauri-build.workspace = true
